+++ JotaNunes.Domain/Interfaces/IEmpreendimentoBaseRepository.cs
@@ -6,5 +6,5 @@ namespace JotaNunes.Domain.Interfaces;
 public interface IEmpreendimentoBaseRepository : IBaseRepository<EmpreendimentoBase>
 {
     public Task<EmpreendimentoBase?> GetByIdAsync(Guid id);
-    public Task<EmpreendimentoBase?> GetByVersionAsync(Guid id, long versao);
+    public Task<EmpreendimentoBase?> GetByVersionAsync(Guid id, int versao);
 }
